BusquedaMateriales

+++ src/main/java/com/example/Gestion/Controladores/MaterialesControlador.java
@@ -1,5 +1,8 @@
 package com.example.Gestion.Controladores;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -12,11 +15,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.example.Gestion.Entidades.Materiales.Materiales;
 import com.example.Gestion.Entidades.Materiales.MaterialesRepositorio;
 import com.example.Gestion.Entidades.TipoMaterial.Tipo_materialRepositorio;
 import com.example.Gestion.Entidades.Usuarios.UsuarioLog;
+import com.example.Gestion.Entidades.Usuarios.Usuarios;
 import com.example.Gestion.Entidades.ClienteProveedor.Cliente_proveedorRepositorio;
 
 @Controller
@@ -27,6 +32,8 @@ public class MaterialesControlador {
     @Autowired Cliente_proveedorRepositorio cliente_proveedorRepositorio;
     @Autowired Tipo_materialRepositorio tipo_materialRepositorio;
     @Autowired UsuarioLog usuarioLog;
+
+    //List<Materiales> Listamateriales = new ArrayList<>();
 
     @GetMapping("/{pagina}")
     public String listaMateriales(@PathVariable("pagina") int pagina, @RequestParam(defaultValue = "10") int tamaño,   
@@ -42,8 +49,16 @@ public class MaterialesControlador {
         return "Materiales/ListaMateriales";
     }
 
+    @GetMapping("/Buscar")
+    @ResponseBody
+    public List<Materiales> buscarMateriales(@RequestParam String nombre, Model model){
+        return materialesRepositorio.findByNombreAndUsuarios(nombre, usuarioLog.correoUsuario());
+    }
+
     @GetMapping("/NuevoMaterial")
+    @ResponseBody
     public String nuevoMaterial(Model model){
+        
         model.addAttribute("Proveedoress", cliente_proveedorRepositorio.findByUsuarios(usuarioLog.correoUsuario(),true));
         model.addAttribute("TiposMaterialess", tipo_materialRepositorio.findAll());
         model.addAttribute("Materialess", new Materiales());
@@ -59,16 +74,29 @@ public class MaterialesControlador {
 
     @GetMapping("/EditarMaterial/{id}")
     public String editarMaterial(@PathVariable("id")int id, Model model){
-        model.addAttribute("Materialess", materialesRepositorio.findById(id).get());
-        model.addAttribute("Proveedoress", cliente_proveedorRepositorio.findByUsuarios(usuarioLog.correoUsuario(),true));
-        model.addAttribute("TiposMaterialess", tipo_materialRepositorio.findAll());
-        return "Materiales/NuevoMaterial";
-
+        Materiales material = materialesRepositorio.findById(id).get();
+        Usuarios ususario = usuarioLog.correoUsuario();
+        if(ususario == material.getUsuarios()){
+            model.addAttribute("Materialess",material);
+            model.addAttribute("Proveedoress", cliente_proveedorRepositorio.findByUsuarios(ususario,true));
+            model.addAttribute("TiposMaterialess", tipo_materialRepositorio.findAll());
+            return "Materiales/NuevoMaterial";
+        }else{
+            model.addAttribute("error", "Material no encontrado...");
+            return "error";
+        }
     }
 
     @GetMapping("/EliminarMaterial/{id}")
-    public String eliminarMaterial(@PathVariable("id")int id){
-        materialesRepositorio.deleteById(id);
-        return "redirect:/GestionZapaterias/Materiales/0";
+    public String eliminarMaterial(@PathVariable("id")int id, Model model){
+        Materiales material = materialesRepositorio.findById(id).get();
+        Usuarios ususario = usuarioLog.correoUsuario();
+        if(ususario == material.getUsuarios()){
+            materialesRepositorio.deleteById(id);
+            return "redirect:/GestionZapaterias/Materiales/0";
+        }else{
+            model.addAttribute("error", "Material no encontrado...");
+            return "error";
+        }
     }
 }
